LPS-166009 Move BaseTemplate, BaseTemplateManager, TemplateContextHelper to portal-template-engine-api
