+++ modules/apps/portal-template/portal-template-engine-api/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Portal Template Engine API
+Bundle-SymbolicName: com.liferay.portal.template.engine.api
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.portal.template.engine
