+++ modules/apps/portal-template/portal-template-engine-api/src/main/java/com/liferay/portal/template/engine/BaseTemplate.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.template;
+package com.liferay.portal.template.engine;
 
 import com.liferay.petra.io.unsync.UnsyncStringWriter;
 import com.liferay.petra.string.StringBundler;
