+++ modules/apps/portal-template/portal-template-engine-api/src/main/java/com/liferay/portal/template/engine/BaseTemplateManager.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.template;
+package com.liferay.portal.template.engine;
 
 import com.liferay.portal.kernel.template.Template;
 import com.liferay.portal.kernel.template.TemplateManager;
