+++ modules/apps/portal-template/portal-template-engine-api/src/main/resources/com/liferay/portal/template/engine/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
